Hayson/TFT/requirements.txt
# Core data processing
pandas>=1.5.0,<2.4.0
numpy>=1.20.0,<2.0.0

# Stock data and corporate actions
yfinance>=0.2.0

# News and embeddings
requests>=2.28.0
transformers>=4.20.0
torch>=1.12.0
sentencepiece>=0.1.97

# Environment variables
python-dotenv>=0.19.0

# Technical indicators (optional - choose one)
# Option 1: Install ta-lib manually first, then uncomment:
# ta-lib>=0.4.0
# Option 2: Use pandas-ta as alternative (no C dependencies):
pandas-ta>=0.3.14b0

# TFT and forecasting
pytorch-forecasting>=0.10.0
lightning>=2.0.0

# Utilities
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0

fredapi>=0.4.0
